+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/qualitynotification/domain/investigation/model/exception/NotificationStatusTransitionNotAllowed.java
@@ -26,6 +26,6 @@ import org.eclipse.tractusx.traceability.qualitynotification.domain.model.Qualit
 public class NotificationStatusTransitionNotAllowed extends RuntimeException {
 
     public NotificationStatusTransitionNotAllowed(String notificationId, QualityNotificationStatus from, QualityNotificationStatus to) {
-        super("Transition from status %s to status %s  is not allowed for investigation with id %s".formatted(from.name(), to.name(), notificationId));
+        super("Transition from status '%s' to status '%s' is not allowed for investigation with id '%s'".formatted(from.name(), to.name(), notificationId));
     }
 }