+++ CHANGELOG.md
@@ -8,9 +8,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
+- Added translationContext "commonAlert" to all modal components
+- Send quality alert to customer / action on queued items
 - Adapt asset synchronization from IRS
 - Cors configuration for environment e2e-a and e2e-b
 - GitHubAction to update dependencies files on merge to main
+- Added testdata with batches
 
 ### Changed
 
@@ -24,6 +27,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Removed
 
 - Bpn Mapping fallback in INT environment profile - has been removed
+
 
 ## [3.5.0] - 2023-05-30